Let TreeType have fixed size.
Note: 1. pos becomes uint32_t for larger compatibility of index
	2^31~2^32-1 in 64-bit systems.
      2. freq remains int32_t to preserve "subtract" implementation in
	comparison functions.

+++ include/internal/chewing-private.h
@@ -77,11 +77,11 @@ typedef struct {
 	uint32_t key;
 	union{
 		struct{
-			int begin, end;
+			int32_t begin, end;
 		} child;
 		struct {
-			long pos;
-			int freq;
+			uint32_t pos;
+			int32_t freq;
 		} phrase;
 	};
 } TreeType;
